Add complex32 dtype support to CPU/GPU implementation of
ghstack-source-id: b544c3608345ebdbb2f39ff193be6fa7fa1b5ec3
Pull Request resolved: https://github.com/pytorch/pytorch/pull/45339

+++ torch/_tensor_str.py
@@ -98,7 +98,8 @@ class _Formatter(object):
             nonzero_finite_max = nonzero_finite_abs.max().double()
 
             for value in nonzero_finite_vals:
-                if value != torch.ceil(value):
+                # Convert to double to add support for Half/ComplexHalf since there’s no ceil() for fp16 on CPU
+                if value != torch.ceil(value.double()):
                     self.int_mode = False
                     break
